chore(graphql): add connectors-common-library to unify http behavior
