+++ connectors/graphql/connector.yml
@@ -1,3 +1,3 @@
 name: GRAPHQL
 type: io.camunda:connector-graphql:1
-variables: [ url, method, authentication, variables, query, connectionTimeoutInSeconds ]
+variables: [ graphql, authentication ]
