+++ connectors/graphql/pom.xml
@@ -55,6 +55,13 @@
       <groupId>org.slf4j</groupId>
       <artifactId>jcl-over-slf4j</artifactId>
     </dependency>
+
+    <!-- TODO : remove version later -->
+    <dependency>
+      <groupId>io.camunda.connector</groupId>
+      <artifactId>connectors-common-library</artifactId>
+      <version>0.15.0-SNAPSHOT</version>
+    </dependency>
   </dependencies>
 
 </project>
